Basically have accessor finding workig.  As long as the class is finalized and the mop is powerful enough, we should be able to find the proper (SETF SLOT-VALUE) form for any given slot reader.
In order for this to work, we are treating this as the default case.  If a Modf
function/method is not defined, we assume we are dealing with a class accessor.

This might break the previous work I did making modf accept recursive builder
functions.

I also have to treat the default container argument as 1.  This is not always
true...

Still doesn't work as it produces actual function objects in the expansion.
This means that several prominent Lisps can't compile files with Modf
expressions anymore.  This will have to be fixed.